`input` test utility consolidation
