+++ src/test-utilities/input/click-fullscreen-control.ts
